fix(adapters/qqguild): unable send figure message
